chat-infrastructure/rag_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
qdrant-client==1.9.2
redis==5.0.3
aiohttp==3.9.3
httpx==0.27.0
